+++ src/lib-main.js
@@ -1,6 +1,8 @@
 import ReactiveMenuItem from './ReactiveMenuItem.vue'
 import ReactiveMenuTree from './ReactiveMenuTree.vue'
+import useReactiveMenu from './useReactiveMenu'
 export {
     ReactiveMenuItem,
-    ReactiveMenuTree
+    ReactiveMenuTree,
+    useReactiveMenu
 }