+++ vite.config.js
@@ -3,10 +3,11 @@ import { fileURLToPath, URL } from 'node:url'
 import { defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
 import vueJsx from '@vitejs/plugin-vue-jsx'
+import cssInjectedByJsPlugin from 'vite-plugin-css-injected-by-js'
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [vue(), vueJsx()],
+  plugins: [vue(), vueJsx(), cssInjectedByJsPlugin()],
   resolve: {
     alias: {
       '@': fileURLToPath(new URL('./src', import.meta.url))
@@ -14,11 +15,13 @@ export default defineConfig({
   },
   build: {
     rollupOptions: {
-      external: ['lodash', 'vue', 'element-plus'],
+      external: ['lodash', 'vue', 'element-plus', 'vue-router'],
       output: {
+        manualChunks: undefined,
         globals: {
           lodash: '_',
           vue: 'Vue',
+          'vue-router': 'VueRouter',
           'element-plus': 'ElementPlus'
         }
       }
@@ -26,7 +29,7 @@ export default defineConfig({
     cssCodeSplit: true,
     lib: {
       entry: './src/lib-main.js',
-      formats: ['umd'],
+      formats: ['es'],
       name: 'ReactiveMenu',
       fileName: 'reactive-menu'
     }